moved valid_codename to source_app/utils.py
