+++ securedrop/source.py
@@ -22,7 +22,7 @@ import store
 from db import db_session, Source, Submission, Reply, get_one_or_else
 from source_app import create_app
 from source_app.decorators import login_required, ignore_static
-from source_app.utils import logged_in
+from source_app.utils import logged_in, valid_codename
 
 import logging
 # This module's logger is explicitly labeled so the correct logger is used,
@@ -322,25 +322,6 @@ def batch_delete():
 
     flash(gettext("All replies have been deleted"), "notification")
     return redirect(url_for('lookup'))
-
-
-def valid_codename(codename):
-    # Ignore codenames that are too long to avoid DoS
-    if len(codename) > Source.MAX_CODENAME_LEN:
-        app.logger.info(
-                "Ignored attempted login because the codename was too long.")
-        return False
-
-    try:
-        filesystem_id = crypto_util.hash_codename(codename)
-    except crypto_util.CryptoException as e:
-        app.logger.info(
-                "Could not compute filesystem ID for codename '{}': {}".format(
-                    codename, e))
-        abort(500)
-
-    source = Source.query.filter_by(filesystem_id=filesystem_id).first()
-    return source is not None
 
 
 @app.route('/login', methods=('GET', 'POST'))
